+++ examples/ferris_hybrid_demo.rs
@@ -199,3 +199,5 @@ fn run_demo() -> Result<()> {
 
     Ok(())
 }
+
+