+++ examples/rmsnorm_bench.rs
@@ -117,3 +117,5 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     Ok(())
 }
+
+