+++ examples/softmax_bench.rs
@@ -107,3 +107,5 @@ fn main() -> Result<(), Box<dyn std::error::Error>> {
 
     Ok(())
 }
+
+