+++ scripts/update-metrics.sh
@@ -75,3 +75,5 @@ echo "  sed -i '' 's/$README_COV/${COVERAGE}%/' README.md"
 
 
 
+
+